backend/requirements.txt
# Core
fastapi>=0.100.0
uvicorn>=0.22.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# Auth
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0,<4.1.0
passlib[bcrypt]>=1.7.4

# AI/ML
crewai>=1.4.0
crewai_tools
openai>=1.0.0
langchain>=0.1.0

# Utils
pydantic>=2.0.0
pydantic[email]
pyyaml>=6.0.0
requests>=2.31.0
google-search-results
psycopg2-binary>=2.9.6
python-multipart>=0.0.5